# Conflicts:
#	Assets/Prefabs/LocalPlayer.prefab
#	Assets/_Scriptables/Spells/DashSpell1.asset

+++ Assets/_Scripts/Gameplay/Modules/PlayerModule.cs
@@ -172,7 +172,7 @@ public class DamagesInfos
 [System.Serializable]
 public class DamagesParameters
 {
-	public uint damageHealth;
+	public ushort damageHealth;
 }
